Move InterpolatePaintScale in display settings package.
