+++ src/main/java/fiji/plugin/trackmate/gui/displaysettings/DisplaySettingsIO.java
@@ -11,8 +11,6 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.stream.Collectors;
 
-import org.jfree.chart.renderer.InterpolatePaintScale;
-
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonDeserializationContext;
