+++ src/main/java/fiji/plugin/trackmate/visualization/PerEdgeFeatureColorGenerator.java
@@ -2,10 +2,10 @@ package fiji.plugin.trackmate.visualization;
 
 import java.awt.Color;
 
-import org.jfree.chart.renderer.InterpolatePaintScale;
 import org.jgrapht.graph.DefaultWeightedEdge;
 
 import fiji.plugin.trackmate.Model;
+import fiji.plugin.trackmate.gui.displaysettings.InterpolatePaintScale;
 
 public class PerEdgeFeatureColorGenerator implements TrackColorGenerator
 {
