+++ src/main/java/fiji/plugin/trackmate/visualization/PerSpotFeatureColorGenerator.java
@@ -2,11 +2,11 @@ package fiji.plugin.trackmate.visualization;
 
 import java.awt.Color;
 
-import org.jfree.chart.renderer.InterpolatePaintScale;
 import org.jgrapht.graph.DefaultWeightedEdge;
 
 import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.Spot;
+import fiji.plugin.trackmate.gui.displaysettings.InterpolatePaintScale;
 
 public class PerSpotFeatureColorGenerator implements TrackColorGenerator
 {
