+++ src/main/java/fiji/plugin/trackmate/visualization/SpotColorGenerator.java
@@ -2,9 +2,8 @@ package fiji.plugin.trackmate.visualization;
 
 import java.awt.Color;
 
-import org.jfree.chart.renderer.InterpolatePaintScale;
-
 import fiji.plugin.trackmate.Spot;
+import fiji.plugin.trackmate.gui.displaysettings.InterpolatePaintScale;
 
 public class SpotColorGenerator implements FeatureColorGenerator< Spot >
 {
