+++ src/main/java/fiji/plugin/trackmate/visualization/SpotColorGeneratorPerEdgeFeature.java
@@ -3,11 +3,11 @@ package fiji.plugin.trackmate.visualization;
 import java.awt.Color;
 import java.util.Set;
 
-import org.jfree.chart.renderer.InterpolatePaintScale;
 import org.jgrapht.graph.DefaultWeightedEdge;
 
 import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.Spot;
+import fiji.plugin.trackmate.gui.displaysettings.InterpolatePaintScale;
 
 public class SpotColorGeneratorPerEdgeFeature implements FeatureColorGenerator< Spot >
 {
